+++ dev-packages/e2e-tests/test-applications/supabase-nextjs/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { supabase } from '@/lib/initSupabase'
+import { supabase } from '@/lib/initSupabaseAnon'
 import { SessionContextProvider } from '@supabase/auth-helpers-react'
 import type { AppProps } from 'next/app'
